+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/CheckListItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CheckListItemRepository extends CrudRepository<CheckListItemEntity, Integer> {
+public interface CheckListItemRepository extends CrudRepository<CheckListItemEntity, Long> {
 }
