+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/CheckListRespostaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CheckListRespostaRepository extends CrudRepository<CheckListRespostaEntity, Integer> {
+public interface CheckListRespostaRepository extends CrudRepository<CheckListRespostaEntity, Long> {
 }
