+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/EmpresaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmpresaRepository extends CrudRepository<EmpresaEntity, Integer> {
+public interface EmpresaRepository extends CrudRepository<EmpresaEntity, Long> {
 }
