+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/HidrometroLocalArmazenamentoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HidrometroLocalArmazenamentoRepository extends CrudRepository<HidrometroLocalArmazenamentoEntity, Integer> {
+public interface HidrometroLocalArmazenamentoRepository extends CrudRepository<HidrometroLocalArmazenamentoEntity, Long> {
 }
