+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/HidrometroTipoInstalacaoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HidrometroTipoInstalacaoRepository extends CrudRepository<HidrometroTipoInstalacaoEntity, Integer> {
+public interface HidrometroTipoInstalacaoRepository extends CrudRepository<HidrometroTipoInstalacaoEntity, Long> {
 }
