+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/ImovelRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImovelRepository extends CrudRepository<ImovelEntity, Integer> {
+public interface ImovelRepository extends CrudRepository<ImovelEntity, Long> {
 }
