+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/MaterialRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaterialRepository extends CrudRepository<MaterialEntity, Integer> {
+public interface MaterialRepository extends CrudRepository<MaterialEntity, Long> {
 }
