+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/MotivoCorteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MotivoCorteRepository extends CrudRepository<MotivoCorteEntity, Integer> {
+public interface MotivoCorteRepository extends CrudRepository<MotivoCorteEntity, Long> {
 }
