+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/MotivoEncerramentoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MotivoEncerramentoRepository extends CrudRepository<MotivoEncerramentoEntity, Integer> {
+public interface MotivoEncerramentoRepository extends CrudRepository<MotivoEncerramentoEntity, Long> {
 }
