+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/OrdemServicoFotoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrdemServicoFotoRepository extends CrudRepository<OrdemServicoFotoEntity, Integer> {
+public interface OrdemServicoFotoRepository extends CrudRepository<OrdemServicoFotoEntity, Long> {
 }
