+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/OrdemServicoSubstituicaoHidrometroRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrdemServicoSubstituicaoHidrometroRepository extends CrudRepository<OrdemServicoSubstituicaoHidrometroEntity, Integer> {
+public interface OrdemServicoSubstituicaoHidrometroRepository extends CrudRepository<OrdemServicoSubstituicaoHidrometroEntity, Long> {
 }
