+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/PavimentoTipoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PavimentoTipoRepository extends CrudRepository<PavimentoTipoEntity, Integer> {
+public interface PavimentoTipoRepository extends CrudRepository<PavimentoTipoEntity, Long> {
 }
