+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/ReligacaoTipoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReligacaoTipoRepository extends CrudRepository<ReligacaoTipoEntity, Integer> {
+public interface ReligacaoTipoRepository extends CrudRepository<ReligacaoTipoEntity, Long> {
 }
