+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/UnidadeMedidaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UnidadeMedidaRepository extends CrudRepository<UnidadeMedidaEntity, Integer> {
+public interface UnidadeMedidaRepository extends CrudRepository<UnidadeMedidaEntity, Long> {
 }
