+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/VeiculoMarcaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VeiculoMarcaRepository extends CrudRepository<VeiculoMarcaEntity, Integer> {
+public interface VeiculoMarcaRepository extends CrudRepository<VeiculoMarcaEntity, Long> {
 }
