+++ mobileos-cloud/mobileos-services/src/main/java/br/com/wsilva/mobileos/repository/VeiculoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VeiculoRepository extends CrudRepository<VeiculoEntity, Integer> {
+public interface VeiculoRepository extends CrudRepository<VeiculoEntity, Long> {
 }
